creating an applicant controller and view, and making a post request to backend ... next steps are to figure out how to extract that data
