+++ app/controllers/applicants_controller.rb
@@ -1,0 +1,12 @@
+class ApplicantsController < ApplicationController
+  protect_from_forgery with: :null_session
+
+  def new
+  end
+
+  def create
+    # Figure out how to spit out body!
+    render plain: params[:firstName].inspect
+  end
+
+end
