+++ app/helpers/applicants_helper.rb
@@ -1,0 +1,2 @@
+module ApplicantsHelper
+end
