+++ app/views/applicants/new.html.erb
@@ -1,0 +1,2 @@
+<h1> Application Form </h1>
+<%= javascript_pack_tag 'form' %>
