+++ config/routes.rb
@@ -1,4 +1,8 @@
 Rails.application.routes.draw do
   # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
   get "/hello_react", to: "hello_react#index"
+
+  # Declares standard REST resource
+  resources :applicants
+  
 end
